CAUSEWAY-4036: min/max fractional digits for all floating point types
