+++ core/metamodel/src/main/java/org/apache/causeway/core/metamodel/valuesemantics/BigDecimalValueSemantics.java
@@ -33,16 +33,12 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
 import org.apache.causeway.applib.services.bookmark.IdStringifier;
-import org.apache.causeway.applib.value.semantics.DefaultsProvider;
 import org.apache.causeway.applib.value.semantics.NumericValueSemantics;
-import org.apache.causeway.applib.value.semantics.Parser;
 import org.apache.causeway.applib.value.semantics.ValueDecomposition;
 import org.apache.causeway.applib.value.semantics.ValueSemanticsProvider;
 import org.apache.causeway.commons.collections.Can;
 import org.apache.causeway.commons.internal.base._Strings;
 import org.apache.causeway.core.config.CausewayConfiguration;
-import org.apache.causeway.core.metamodel.context.MetaModelContext;
-import org.apache.causeway.core.metamodel.util.Facets;
 import org.apache.causeway.schema.common.v2.ValueType;
 import org.apache.causeway.schema.common.v2.ValueWithTypeDto;
 
@@ -53,11 +49,8 @@ import lombok.Setter;
 @Primary
 //has no effect @Priority(PriorityPrecedence.LATE)
 public class BigDecimalValueSemantics
-extends NumericValueSemantics<BigDecimal>
-implements
-    DefaultsProvider<BigDecimal>,
-    Parser<BigDecimal>,
-    IdStringifier.EntityAgnostic<BigDecimal> {
+extends NumericValueSemanticsAbstract<BigDecimal>
+implements IdStringifier.EntityAgnostic<BigDecimal> {
 
     @Setter @Inject
     private CausewayConfiguration causewayConfiguration;
@@ -75,6 +68,11 @@ implements
     @Override
     public BigDecimal getDefaultValue() {
         return BigDecimal.ZERO;
+    }
+
+    @Override
+    protected boolean isIntegerOnly() {
+        return false;
     }
 
     // -- COMPOSER
@@ -119,36 +117,20 @@ implements
     }
 
     @Override
-    public void configureDecimalFormat(
+    protected void configureDecimalFormat(
             final ValueSemanticsProvider.Context context, final DecimalFormat format, final FormatUsageFor usedFor) {
-        if(context==null)
-            return;
 
-        var feature = MetaModelContext.instanceElseFail().getSpecificationLoader()
-            .loadFeature(context.featureIdentifier())
-            .orElse(null);
-        if(feature==null)
-            return;
-
-        // evaluate any facets that provide the MaximumFractionDigits
-        Facets.maxFractionalDigits(feature)
-                .ifPresent(format::setMaximumFractionDigits);
+        super.configureDecimalFormat(context, format, usedFor);
 
         var bigDecimalConfig = causewayConfiguration.valueTypes().bigDecimal();
         if(!usedFor.isParsing()
                 || bigDecimalConfig.editing().preserveScale()) {
-
-            // if there is a facet specifying minFractionalDigits (ie the scale), then apply it
-            Facets.minFractionalDigits(feature)
-            .ifPresentOrElse(
-                    format::setMinimumFractionDigits,
-                    ()->{
-                        // otherwise, apply a minScale if configured.
-                        Optional.ofNullable(bigDecimalConfig.display().minScale())
-                        .ifPresent(format::setMinimumFractionDigits);
-                    });
+            if(format.getMinimumFractionDigits()==0) {
+                // apply a minScale if configured.
+                Optional.ofNullable(bigDecimalConfig.display().minScale())
+                    .ifPresent(format::setMinimumFractionDigits);
+            }
         }
-
     }
 
     @Override
@@ -170,7 +152,7 @@ implements
     @Component
     @Qualifier(NumericValueSemantics.NO_GROUPING)
     public static class NoGrouping extends BigDecimalValueSemantics {
-        @Override protected GroupingSeparatorProvider grouping() {
+        @Override public GroupingSeparatorProvider grouping() {
             return GroupingSeparatorProvider.NO_GROUPING;
         }
     }
@@ -178,7 +160,7 @@ implements
     @Component
     @Qualifier(NumericValueSemantics.LOCALE_GROUPING_DISPLAY)
     public static class LocaleGroupingDisplay extends BigDecimalValueSemantics {
-        @Override protected GroupingSeparatorProvider grouping() {
+        @Override public GroupingSeparatorProvider grouping() {
             return GroupingSeparatorProvider.LOCALE_GROUPING_DISPLAY;
         }
     }
@@ -186,7 +168,7 @@ implements
     @Component
     @Qualifier(NumericValueSemantics.LOCALE_GROUPING_ALL)
     public static class LocaleGroupingAll extends BigDecimalValueSemantics {
-        @Override protected GroupingSeparatorProvider grouping() {
+        @Override public GroupingSeparatorProvider grouping() {
             return GroupingSeparatorProvider.LOCALE_GROUPING_ALL;
         }
     }
